⚡️dev/multi-mayQwik5≡ refactored some of the code into a custom hook | pushing PR..
