+++ qwikJsLearning2023/poke-qwik/src/hooks/index.ts
@@ -1,0 +1,7 @@
+// FILE: global `hooks` export file
+// _______________________________________________
+
+export * from "./use-counter"
+export * from "./use-pokemon-game"
+export * from "./use-list-ssr"
+// _______________________________________________
